feat: 음식점 상세 정보 페이지 매핑 메서드 RestaurantViewController -> RestaurantController로 이동
